added save data in reset

+++ reset.py
@@ -2,12 +2,22 @@
 
 import os
 from config.config import CONFIG_PATH, DATABASE_PATH
+from utils.functions import yes_no_prompt
+from export import main as export_habits
 
 
 def main():
     """main function"""
+    ask_reset = yes_no_prompt(
+        "Are you sure you want to reset Habitpy? (y/n)\nWARNING THIS WILL DELETE ALL YOUR DATA\nWE WILL GENERATE A COPY OF YOUR DATA AS A CSV FILE IN ....\n=> "
+    )
+    if "y" not in ask_reset:
+        print("Reset cancelled.")
+        return
+    export_habits()
     os.remove(DATABASE_PATH)
     os.rmdir("user")
     print("removed user folder")
     os.remove(CONFIG_PATH)
     print("removed config.json")
+    print("Habitpy has been reset, please run 'habitpy setup' to start again.")
